Issue #985 Move feedback type to store
