+++ src/application/routing.ts
@@ -5,7 +5,7 @@ import { History, Location } from 'history';
 import { Id as LearnId } from '../stores/explore';
 import { Id as TopicId } from '../stores/topics';
 import * as R from 'ramda';
-import { ServiceFeedback } from '../components/services/use_send_feedback';
+import { ServiceFeedback } from '../stores/feedback/types';
 
 // The property names of this structure are defined by the corresponding
 // route definitions, e.g. parsing a url '/learn/1' which matches
