+++ src/components/feedback/feedback_modal_container.tsx
@@ -12,8 +12,9 @@ import { ServiceDetailIconComponent } from '../services/service_detail_icon';
 import { EmptyComponent } from '../empty_component/empty_component';
 
 import { FeedbackOptionsModalComponent } from './feedback_options_modal_component';
-import { SendFeedbackPromise, ServiceFeedback } from '../services/use_send_feedback';
+import { SendFeedbackPromise } from '../services/use_send_feedback';
 import { showToast } from '../../application/toast';
+import { ServiceFeedback } from '../../stores/feedback/types';
 
 interface FeedbackModalContainerProps {
     readonly feedbackEnabled: boolean;
